most other easy scripts

+++ compile.py
@@ -14,45 +14,49 @@ with open("file_hashes.json", "r", encoding="utf-8") as f:
 os.makedirs("build", exist_ok=True)
 os.chdir("build")
 
-for full_path in glob.glob("../scripts/*.p"):
+paths = glob.glob("../scripts/*.p")
+for i, full_path in enumerate(paths):
     path = Path(full_path)
     script = path.stem
+    src_time = os.path.getmtime(full_path)
 
-    print(f"Compiling {script}.p")
-    subprocess.check_call(
-        [
-            "../gf-pawncc/build/compiler/gf-pawncc",
-            f"../scripts/{script}.p",
-            "-d0",
-            "-i=../scripts/",
-            "-a",
-            "-S1024",
-        ],
-    )
-    subprocess.check_call(
-        [
-            "../gf-pawncc/build/compiler/gf-pawncc",
-            f"../scripts/{script}.p",
-            "-d0",
-            "-i../scripts/",
-            "-S1024",
-        ],
-    )
+    print(f"Compiling {script}.p {i+1}/{len(paths)}")
+    # basic modification check to avoid rebuilding
+    if not os.path.exists(f"{script}.amx") or (os.path.getmtime(f"{script}.amx") < src_time):
+        subprocess.check_call(
+            [
+                "../gf-pawncc/build/compiler/gf-pawncc",
+                f"../scripts/{script}.p",
+                "-d0",
+                "-i=../scripts/",
+                "-a",
+                "-S1024",
+            ],
+        )
+        subprocess.check_call(
+            [
+                "../gf-pawncc/build/compiler/gf-pawncc",
+                f"../scripts/{script}.p",
+                "-d0",
+                "-i../scripts/",
+                "-S1024",
+            ],
+        )
 
-    subprocess.check_call(["mv", f"../scripts/{script}.amx", "."])
+        subprocess.check_call(["mv", f"../scripts/{script}.amx", "."])
 
-    subprocess.check_call(["mv", f"../scripts/{script}.asm", "."])
+        subprocess.check_call(["mv", f"../scripts/{script}.asm", "."])
 
-    disasm = PawnDisassembler(f"{script}.amx")
-    with open(f"{script}.pasm", "w+", encoding="utf-8") as f:
-        f.write(disasm.infodump())
-        f.write(disasm.disasm())
-    disasm.dump_data(f"{script}.data.bin")
+        disasm = PawnDisassembler(f"{script}.amx")
+        with open(f"{script}.pasm", "w+", encoding="utf-8") as f:
+            f.write(disasm.infodump())
+            f.write(disasm.disasm())
+        disasm.dump_data(f"{script}.data.bin")
 
-    subprocess.run(
-        ["git", "diff", "--no-index", f"../disasm/{script}.pasm", f"{script}.pasm"],
-        check=False,
-    )
+        subprocess.run(
+            ["git", "diff", "--no-index", f"../disasm/{script}.pasm", f"{script}.pasm"],
+            check=False,
+        )
 
     with open(f"{script}.amx", "rb") as f:
         FILE_HASH = hashlib.sha1(f.read()).hexdigest()
